+++ programs/amm/src/err.rs
@@ -9,6 +9,8 @@ pub enum AmmError {
     /// about the issue.
     #[msg("Provided account breaks some constraints, see logs for more info")]
     InvalidAccountInput,
+    #[msg("One of the provided input arguments is invalid")]
+    InvalidArg,
     /// Use this error for program paths which should never be reached if the
     /// program logic works as intended.
     #[msg("There's a bug in the program, see logs for more info")]
@@ -19,4 +21,10 @@ pub fn acc(msg: impl Display) -> AmmError {
     msg!("[InvalidAccountInput] {}", msg);
 
     AmmError::InvalidAccountInput
+}
+
+pub fn arg(msg: impl Display) -> AmmError {
+    msg!("[InvalidArg] {}", msg);
+
+    AmmError::InvalidArg
 }